+++ TSClientGen.Contract/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyVersion("7.2.0")]
-[assembly: AssemblyInformationalVersion("7.2.0")]
+[assembly: AssemblyVersion("8.0.0")]
+[assembly: AssemblyInformationalVersion("8.0.0")]
 
 [assembly: AssemblyTitle("TSClientGen.Contract")]
 [assembly: AssemblyDescription("Contracts assembly with attributes for decorating api for TSClientGen tool")]